+++ lib/src/forwarding/forwarders/lsl/device_name/device_name_stub.dart
@@ -1,8 +1,0 @@
-import 'device_name.dart';
-
-LslDeviceNameProvider createLslDeviceNameProvider() => _StubLslDeviceNameProvider();
-
-class _StubLslDeviceNameProvider implements LslDeviceNameProvider {
-  @override
-  String get deviceName => 'Phone';
-}